Don't run the password validation unless the password is changing
In one of my applications I use my devise model for information
beyond just credentials. That means that we will often be updating
details without changing the password.

Previously, that would throw an exception of

```undefined method `downcase' for nil:NilClass```

because the validation would attempt to run against a value that
wasn't set.

Skipping the validation if no password is given shouldn't cause any
problems as devise will validate for password presence when necessary.